testwithdbt/requirements.txt
dbt-postgres>=1.6.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
pyarrow>=12.0.0
